Weapons bug fixed
I added a `weapons` key in the user's `backpack`, and set the value of it as a dict. Inside the `weapons` there is a nested dict called `weapons`, which is to hold all the weapons a user has. I forgot to do `["weapons"]["weapons"]` when counting all the weapons in the backpack. Bug is fixed.